extract rental from engine
