+++ rental.js
@@ -1,0 +1,6 @@
+export class Rental {
+    constructor(movie, daysRented) {
+        this.movie = movie
+        this.daysRented = daysRented
+    }
+}
